Break up srdf_model.cpp into smaller files and fix requested changes
