+++ tesseract_python/tesseract_python/swig/tesseract_python.i
@@ -87,6 +87,7 @@
 %include "tesseract_geometry/geometry_loaders.i"
 %include "tesseract_scene_graph/graph.i"
 %include "tesseract_scene_graph/resource_locator.i"
+%include "tesseract_scene_graph/srdf/types.i"
 %include "tesseract_scene_graph/srdf_model.i"
 %include "tesseract_kinematics/core/forward_kinematics.i"
 %include "tesseract_kinematics/core/forward_kinematics_factory.i"
